1. Oppdatert nvmrc til 24 ( 24.14.0) siste lts. 2. fikset env import i Serviceworker.js bruker nå import.meta.env. 3. Flyttet innholdet fra index.jsx til main.jsx, istedenfor at main.jsx bare er en wrapper. 4. Fikset testing, Enzyme react testing var påbegynt og ikke fullført, skapte stopp i testing.
