+++ .nvmrc
@@ -1,1 +1,1 @@
-v12.13.1
+24
