+++ index.html
@@ -91,7 +91,7 @@
   <title>Kartkatalogen</title>
 </head>
 
-<script type="module" src="/src/main.jsx"></script>
+
 <body>
   <!-- Google Tag Manager (noscript) -->
   <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-MR2X5P"
@@ -106,6 +106,8 @@
     }
   </script>
   <div id="root"></div>
+
+  <script type="module" src="/src/main.jsx"></script>
 </body>
 
 </html>
