+++ jest.config.js
@@ -6,9 +6,11 @@ module.exports = {
   setupFilesAfterEnv: ["<rootDir>/src/setupTests.js"],
   moduleDirectories: ['node_modules', 'src'],
   moduleNameMapper: {
-    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-    "\\.(css|scss|less)$": "identity-obj-proxy"
-  },
+  "^@/(.*)$": "<rootDir>/src/$1",
+
+  "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+  "\\.(css|scss|less)$": "identity-obj-proxy"
+},
   testResultsProcessor: "jest-teamcity-reporter",
   coverageReporters: ["lcov", "text", "teamcity"]
 };
