+++ src/App.test.disabled.jsx
@@ -17,10 +17,15 @@ function setup() {
   }
 }*/
 
-describe('App', () => {
+/*describe('App', () => {
     it('should render self and subcomponents', () => {
       //const { wrapper } = setup()
       //expect(wrapper).toMatchSnapshot();
       expect(1).toBe(1);
     })
 })
+
+*/
+
+//Kommenterte ut resten også, den gjør ingenting egentlig. 
+//Endret navn slik at den også ikke kjøres i test.
